+++ benches/benchmarks.rs
@@ -13,8 +13,10 @@ fn end_to_end_benchmark() {
         log_level: logger::LogLevel::Debug,
         markup_types: vec![MarkupType::Markdown],
         ignore_links: vec![],
+        match_file_extension: false,
         ignore_path: vec![],
         root_dir: None,
+        throttle: 0,
         no_web_links: false,
     };
     let _ = mlc::run(&config);