+++ tests/end_to_end.rs
@@ -19,6 +19,8 @@ async fn end_to_end() {
         log_level: logger::LogLevel::Debug,
         markup_types: vec![MarkupType::Markdown],
         no_web_links: false,
+        match_file_extension: false,
+        throttle: 0,
         ignore_links: vec![],
         ignore_path: vec![
             fs::canonicalize("benches/benchmark/markdown/ignore_me.md").unwrap(),
@@ -45,8 +47,10 @@ async fn end_to_end_different_root() {
         log_level: logger::LogLevel::Debug,
         markup_types: vec![MarkupType::Markdown],
         no_web_links: false,
+        match_file_extension: false,
         ignore_links: vec![],
         ignore_path: vec![],
+        throttle: 0,
         root_dir: Some(test_files),
     };
     if let Err(e) = mlc::run(&config).await {